[NEW][Examples] [simulations.senatorassembly -> to swarms.sim]
