+++ examples/api_examples/legal_team.py
