+++ swarms/structs/ma_utils.py
@@ -99,14 +99,14 @@ models = [
     "anthropic/claude-3-sonnet-20240229",
     "openai/gpt-4o-mini",
     "openai/gpt-4o",
-    "deepseek/deepseek-chat",
-    "deepseek/deepseek-reasoner",
     "groq/deepseek-r1-distill-qwen-32b",
     "groq/deepseek-r1-distill-qwen-32b",
     # "gemini/gemini-pro",
     # "gemini/gemini-1.5-pro",
+    "groq/moonshotai/kimi-k2-instruct",
     "openai/03-mini",
     "o4-mini",
+    "claude-sonnet-4-20250514",
     "o3",
     "gpt-4.1",
     "groq/llama-3.1-8b-instant",
